+++ src/auth.php
@@ -16,6 +16,7 @@
     <meta content="<?php echo $domain ?>/images/logo.png" property="og:image" />
     <meta content="<?php echo $vtccolor ?>" data-react-helmet="true" name="theme-color" />
     <meta content="<?php echo $domain ?>/images/bg.jpg" name="twitter:card">
+    <link rel="stylesheet" href="/css/font.css">
 
 <style>
     .formRight form section button {
@@ -149,7 +150,7 @@
                 },
                 success: function (data) {
                     if (data.error == false) {
-                        $("#title").html("Welcome Back");
+                        $("#title").html(<?php echo $vtcname ?>);
                         $("#msg").html("You are being redirected to Drivers Hub.");
                         window.location.href = "/";
                     } else {